[Graph Coloring] Add Kata Implementation
This commit implements a Kata exploring using Grover's search to solve
the constraint satisfaction problem of graph coloring, coloring nodes of
a graph in such a way that no adjacent nodes share the same color.

This project was done as part of UW CSE 490 Q as its final project.
The authors are Dan Tran, Daniel Snitkovskiy, and Frederik Schmitt.