Use fail instead of returning default values
With the classical Alice and Bob tasks, they have to return a value for
it to compile, so the way other katas do it is to return some default
value before the user completes the task. But if they both return false
for example, then that's already the optimal strategy. :) They could
return different values to fail the tests but that gives away the
solution a little. Using fail statements gets around this.