Let the kata notebooks override all attempted tasks
In https://github.com/microsoft/QuantumKatas/pull/115 we observed that a kata notebook only overrides the task that is currently tested, and uses stubs from the project for the rest of the tasks. This prevents us from implementing tasks which require two related operations be implemented at once, or ones which rely on previously completed tasks.

This change adds tracking of previously attempted tasks and overriding them together with the currently attempted task.