[MultiQubitGates] Clarify controlled gates behavior (#479)
As pointed out in #379, the phrasing "the controlled operation should not modify the state of the control qubits" can be very confusing to a learner, especially when dealing with entanglement and phase kickback later. I tried to explain the need to uncompute the transformation we did to the control qubit and to do a detailed walkthrough of one example.