Add Test attribute to open-ended tasks to show them in VS (#540)
In #478 we removed TestSuiteRunner files that detected tests based on a fixed suffix, but we missed the open-ended tasks in Tasks.qs which should also be defined as tests so that they would appear in the list of tests in VS/VS Code. This change adds @Test attribute to those tasks.