Update measurement counting logic (#592)
With the feature work from https://github.com/microsoft/qsharp-runtime/issues/249 merged, the beta packages will now report measurement counts differently. This is expected, as `M` calls into `Measure`, so counting both operations will double-count measurements and fail validation of solutions. This change adjusts to only count `Measure` in order to ensure the right behavior in the pipeline and when the feature makes it into a public release.

This fixes Measurement counting mechanism should be updated for target package feature. #552